:wrench: Minor changes

+++ src/plugins/CriticalPointExtraction/VestecCriticalPointExtractionAlgorithm.h
@@ -103,7 +103,7 @@ class CriticalPointExtractor {
      * currentSingularity: The singularity e.g. zero vector (0,0,0)
      * vecMatrix: The matrix used to compute the determinant
      */
-    double ComputeDeterminant(std::array<vtkIdType, 4> &tmpIds, DynamicMatrix &vecMatrix, bool usePoints, short perturbationID = -1);
+    double ComputeDeterminant(std::array<vtkIdType, 4> &tmpIds, DynamicMatrix &vecMatrix, bool usePoints, int perturbationID = -1);
    
     /**
      * Check if the singularity is in cell
